requirements.txt
biopython==1.80
numpy==1.24.2
pysam==0.20.0
